mgr/dashboard_v2: Simplify `CephService` class (squash with ee45dabc512e77252065b4624c4f3a0f724ef89a)
This commit will simplify the `CephService` class by removing the meta
class `ServiceMeta` together with the `Service` class and introducing
`CephService` as a `CephServiceMixin` class.

In addition to the removal of these two classes, the injection of the
mgr instance becomes unnecessary as all controller classes do have their
own reference to the mgr object.

Furthermore, this commit will introduce a Python stub file which
enables IDEs to correctly detect the type of the `mgr` instance within
the newly created mixin.

Signed-off-by: Patrick Nawracay <pnawracay@suse.com>