mgr/dashboard: Error when retrieving OSD data - OSD list broken
Fixes a broken OSD list which occurs due to the data for OSDs being
taken from multiple sources.  If the OSD has just been removed but not
purged, chances are that there is a reference left in the CRUSH map,
causing the data collection process to fail.  The solution here is not
to collect data from OSDs which have been removed, resulting in the
corresponding OSD not being displayed on the OSD list.

This leaves the dashboard in a state from which the user (currently)
cannot recover by using the dashboard. But it's also not possible to get
into this siutation as the `remove` functionality has already been
removed.

Fixes: https://tracker.ceph.com/issues/36086

Signed-off-by: Patrick Nawracay <pnawracay@suse.com>